+++ src/commands.rs
@@ -4,11 +4,21 @@ use color_eyre::{
     eyre::{bail, Context},
     Result,
 };
-use subprocess::{Exec, ExitStatus, Redirection};
+use subprocess::{Exec, ExitStatus, Pipeline, Redirection};
 use thiserror::Error;
 use tracing::{debug, info};
 
-use crate::installable::Installable;
+use crate::{installable::Installable, util::get_current_system};
+
+// for some reason there isnt a common trait that both Exec and Pipeline implements, so just use
+// the no op : from bash
+fn ssh_wrap(cmd: Exec, ssh: Option<&str>) -> Pipeline {
+    if let Some(ssh) = ssh {
+        Exec::cmd("echo").arg(cmd.to_cmdline_lossy().as_str()) | Exec::cmd("ssh").arg("-T").arg(ssh)
+    } else {
+        Exec::cmd(":") | cmd
+    }
+}
 
 #[derive(Debug)]
 pub struct Command {
@@ -102,15 +112,8 @@ impl Command {
         } else {
             Exec::cmd(&self.command).args(&self.args)
         };
-        let cmd = if let Some(ssh) = &self.ssh {
-            Exec::cmd("ssh")
-                .arg("-T")
-                .arg(ssh)
-                .stdin(cmd.to_cmdline_lossy().as_str())
-        } else {
-            cmd
-        };
-        let cmd = cmd.stderr(Redirection::None).stdout(Redirection::None);
+        let cmd =
+            ssh_wrap(cmd.stderr(Redirection::None), self.ssh.as_deref()).stdout(Redirection::None);
 
         if let Some(m) = &self.message {
             info!("{}", m);
@@ -155,6 +158,7 @@ pub struct Build {
     installable: Installable,
     extra_args: Vec<OsString>,
     nom: bool,
+    builder: Option<String>,
 }
 
 impl Build {
@@ -164,6 +168,7 @@ impl Build {
             installable,
             extra_args: vec![],
             nom: false,
+            builder: None,
         }
     }
 
@@ -179,6 +184,11 @@ impl Build {
 
     pub fn nom(mut self, yes: bool) -> Self {
         self.nom = yes;
+        self
+    }
+
+    pub fn builder(mut self, builder: Option<String>) -> Self {
+        self.builder = builder;
         self
     }
 
@@ -206,9 +216,16 @@ impl Build {
                     .arg("build")
                     .args(&installable_args)
                     .args(&["--log-format", "internal-json", "--verbose"])
+                    .args(&match &self.builder {
+                        Some(host) => vec![
+                            "--builders".to_string(),
+                            format!("ssh://{host} {} - - 100", get_current_system().unwrap()),
+                        ],
+                        None => vec![],
+                    })
                     .args(&self.extra_args)
-                    .stdout(Redirection::Pipe)
                     .stderr(Redirection::Merge)
+                    .stdout(Redirection::Pipe)
                     | Exec::cmd("nom").args(&["--json"])
             }
             .stdout(Redirection::None);
@@ -219,8 +236,8 @@ impl Build {
                 .arg("build")
                 .args(&installable_args)
                 .args(&self.extra_args)
-                .stdout(Redirection::None)
-                .stderr(Redirection::Merge);
+                .stderr(Redirection::Merge)
+                .stdout(Redirection::None);
 
             debug!(?cmd);
             cmd.join()